+++ CHANGES.md
@@ -29,6 +29,7 @@ Apollo 3.0.0
 * [Feature: add batch create/update/delete OpenAPI operations for namespace items](https://github.com/apolloconfig/apollo/pull/5665)
 * [Change: upgrade Apollo server baseline to Spring Boot 4.1.1 and Spring Cloud 2025.1.3](https://github.com/apolloconfig/apollo/pull/5671)
 * [Fix: preserve last-modified metadata of unchanged items when revoking changes](https://github.com/apolloconfig/apollo/pull/5672)
+* [Fix: support user-token deletion of keys containing slashes or backslashes](https://github.com/apolloconfig/apollo/pull/5676)
 
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo/milestone/18?closed=1)